requirements.txt
# Main Project Requirements
# Python 3.11+

# Core dependencies
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.67.1
pydantic>=2.11.5
beautifulsoup4>=4.12.3
markitdown>=0.0.1a2

# Core LlamaIndex dependencies  
llama-index==0.13.3
llama-index-core==0.13.3
llama-index-workflows==1.3.0
llama-index-cli==0.5.0

# Supabase and PostgreSQL
psycopg2-binary>=2.9.9
supabase>=2.0.0

# Vertex AI RAG dependencies
google-cloud-aiplatform[adk,agent-engines]>=1.108.0
google-adk>=1.10.0
google-auth>=2.35.0
google-api-python-client>=2.154.0

# LLM providers
llama-index-llms-openai==0.5.4
llama-index-llms-anthropic==0.8.5
openai>=1.81.0,<2.0.0
anthropic==0.62.0

# Google Cloud Storage (for Drive operations)
google-cloud-storage>=2.18.2

# Additional utilities
aiohttp>=3.11.11
httpx>=0.28.1

